Wraps configuration in new DSLContexts (for transactions).
This is the right way to use jOOQ's transactions
(see http://www.jooq.org/doc/3.6/manual/sql-execution/transaction-management/).

Authoring this PR makes me wonder if all the DAO methods should take a context
as a parameter instead of injecting it. We should make it hard to accidentally
use the incorrect DSLContext.